requirements.txt
absl-py==0.13.0
addict==2.4.0
antlr4-python3-runtime==4.8
anyio==3.3.0
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==20.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
black==21.4b2
bleach==4.0.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
chumpy==0.70
click==8.0.1
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
deprecation==2.1.0
detectron2==0.5+cu102
entrypoints==0.3
future==0.18.2
fvcore==0.1.5.post20210812
google-auth==1.35.0
google-auth-oauthlib==0.4.5
graphviz==0.20
grpcio==1.39.0
hydra-core==1.1.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.4
importlib-resources==5.2.2
iopath==0.1.8
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyter-packaging==0.10.4
jupyter-server==1.10.2
jupyterlab==3.1.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.1
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lpips==0.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.3
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.0
open3d==0.13.0
opencv-python==4.5.5.62
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pkg_resources==0.0.0
plotly==5.2.1
portalocker==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.10.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-msssim==0.2.1
pytorch3d==0.5.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.8.8
toml==0.10.2
tomlkit==0.7.2
torchaudio==0.9.0
torchvision==0.10.0
torchviz==0.0.2
tornado==6.1
tqdm==4.62.1
traitlets==5.0.5
transforms3d==0.3.1
trimesh==3.9.28
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
zipp==3.5.0
